requirements.txt
asttokens==2.2.1
backcall==0.2.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
cryptography==39.0.2
debugpy==1.6.6
decorator==5.1.1
Deprecated==1.2.13
executing==1.2.0
firebase==4.0.1
firebase-admin==6.1.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
gcloud==0.18.3
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
greenlet==2.0.2
grpcio==1.51.3
grpcio-status==1.51.3
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
ipykernel==6.21.2
ipython==8.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.0.3
jupyter_core==5.2.0
jwcrypto==1.4.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
msgpack==1.0.4
nest-asyncio==1.5.6
oauth2client==4.1.3
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.1.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.22.0
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
Pygments==2.14.0
PyJWT==2.6.0
pyparsing==3.0.9
pypng==0.20220715.0
Pyrebase4==4.6.0
python-dateutil==2.8.2
python-jwt==4.0.0
pywin32==305
pyzmq==25.0.0
qrcode==7.4.2
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.3
sseclient==0.0.27
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
typing_extensions==4.4.0
uplink==0.9.7
uritemplate==4.1.1
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.2
wrapt==1.15.0
